third element & df_corr input
This third element 'Group' helps users to know which group has been estimated.
Adding of "df_corr" in inputs. df_corr corresponds to the output of the function convert.character. It allows to find the correspondences of the character / factor variables recoded in numeric during the creation of the pointer