Output, input change change in big.model.FElasso
The third output 'Group' helps users to know which group has been estimated.
Adding of "df_corr" in inputs. "df_corr" corresponds to the function convert.character output. It allows to find the correspondences of the character / factor variables recoded in numeric during the creation of the pointer or the data in input of linear modelisation